MVP finished login
