+++ instaclone/src/components/PostContainer/PostHeader.js
@@ -2,10 +2,10 @@ import React from 'react';
 
 const PostHeader = props => {
   return(
-    <div className="post-header">
+    <header className="post-header">
       <img src={props.thumbnail} alt={props.username} />
       <h3>{props.username}</h3>
-    </div>
+    </header>
   )
 }
